+++ R/format_estimates.R
@@ -66,6 +66,7 @@ format_estimates <- function(
       stop('To use the `stars` argument, the `tidy` function must produce a column called "p.value"')
     }
     est$stars <- make_stars(est$p.value, stars)
+    est$stars[is.na(est$stars)] <- ""
   }
   if (isTRUE(is_star) && isFALSE(is_glue)) {
     estimate_glue[1] <- paste0(estimate_glue[1], "{stars}")
@@ -128,8 +129,7 @@ https://vincentarelbundock.github.io/modelsummary', group_name))
     times         = grep("modelsummary_tmp\\d+$", colnames(est), value = TRUE),
     v.names       = "value",
     timevar       = "statistic",
-    direction     = "long",
-    new.row.names = 1:1000)
+    direction     = "long")
   est$id <- NULL
 
   # sort then convert back to character